% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centene_toc_ex.R
\docType{data}
\name{centene_toc_ex}
\alias{centene_toc_ex}
\title{Example of a Centene TOC file}
\format{
data frame containing 2 rows and 8 variables
}
\usage{
centene_toc_ex
}
\description{
Example of a table of contents (TOC) file from
Centene's Transparency in Coverage (TiC) site.
}
\keyword{internal}
