% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defog_inn.R
\name{defog_inn}
\alias{defog_inn}
\title{Download and Parse TiC In-Network JSON urls}
\usage{
defog_inn(url)
}
\arguments{
\item{url}{url from third-party-payer's
machine-readable Transparency-in-Coverage filings,
specifically for In-Network plans}
}
\value{
A tibble containing information regarding
health insurance plans' negotiated rates for
in-network procedures performed by healthcare providers.
}
\description{
Download and Parse TiC In-Network JSON urls
}
\examples{
\dontrun{defog_inn(inn_url_ex)}
}
