% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defog_npi.R
\name{defog_npi}
\alias{defog_npi}
\title{Search the NPPES by NPI}
\usage{
defog_npi(npi)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
Search the NPPES by NPI
}
\examples{
defog_npi(1528060837)
}
