% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defog_npi_wide.R
\name{defog_npi_wide}
\alias{defog_npi_wide}
\title{Tidy NPI NPPES Search Data}
\usage{
defog_npi_wide(df)
}
\arguments{
\item{df}{data.frame or tbl_df, response from prov_nppes_npi()}
}
\value{
A tibble containing the NPI(s) searched for on the
NPPES, the date-time of the search, and the unnested and
tidied list-columns of results split into groups.
}
\description{
Tidy NPI NPPES Search Data
}
\examples{
\dontrun{defog_npi_wide(nppes_npi_response)}
}
