% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defog_oon.R
\name{defog_oon}
\alias{defog_oon}
\title{Download and Parse TiC Out-of-Network JSON urls}
\usage{
defog_oon(url)
}
\arguments{
\item{url}{url from third-party-payer's
machine-readable Transparency-in-Coverage filings,
specifically for Out-of-Network (OON) plans;
also referred to as "Allowed Amounts" in the URLs}
}
\value{
A tibble containing information regarding
health insurance plans' allowed amounts for
out-of-network procedures performed by
healthcare providers.
}
\description{
Download and Parse TiC Out-of-Network JSON urls
}
\examples{
\dontrun{defog_oon(oon_url_ex)}
}
