% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defog_toc.R
\name{defog_toc}
\alias{defog_toc}
\title{Download and Parse TiC TOC urls}
\usage{
defog_toc(url)
}
\arguments{
\item{url}{url from a third-party-administrator's (TPA)
machine-readable Transparency-in-Coverage (TiC) filings
containing the Table of Contents (TOC) of plans.}
}
\value{
A tibble containing the TPA's list of insurance plan
names, plan ID, ID type, market type(individual or group),
the parent entity, rate type (in-network or out-of-network),
and url location of each rate type's file.
}
\description{
Download and Parse TiC TOC urls
}
