% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defog_toc_centene.R
\name{defog_toc_centene}
\alias{defog_toc_centene}
\title{Centene: Add Reporting Entity & the Sub-Entity}
\usage{
defog_toc_centene(df)
}
\arguments{
\item{df}{data frame from Centene TOC link}
}
\value{
A tibble adding the reporting entity's name
and the sub-entity's name.
}
\description{
Centene: Add Reporting Entity & the Sub-Entity
}
\examples{
defog_toc_centene(centene_toc_ex)
}
