% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defogger-package.R
\docType{package}
\name{defogger-package}
\alias{defogger}
\alias{defogger-package}
\title{defogger: Tools for Transparency in Coverage (TiC) Files}
\description{
The goal of defogger is to provide a collection of reliable open-source tools to work with Transparency in Coverage (TiC) files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/defogger}
  \item \url{https://andrewallenbruce.github.io/defogger/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/defogger/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewbruce.himni@gmail.com}

}
\keyword{internal}
