% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_toc_url_ex.R
\docType{data}
\name{tic_toc_url_ex}
\alias{tic_toc_url_ex}
\title{Example of a TOC url}
\format{
A vector containing a URL link
}
\usage{
tic_toc_url_ex
}
\description{
Example url link to a machine readable table of contents
(TOC) file from a Transparency in Coverage (TiC) site.
}
\keyword{internal}
